requirements.txt
SQLAlchemy~=2.0.44
Werkzeug~=3.1.4
Flask-JWT-Extended~=4.6.0
schedule~=1.2.2
ollama~=0.6.1
Flask~=3.0.0
docx~=0.2.4
python-docx~=1.2.0
chromadb~=1.3.5
Flask-Migrate~=4.0.5
python-dotenv~=1.0.0
alembic~=1.13.1